+++ dune/gdt/localevaluation/product.hh
@@ -81,7 +81,7 @@ public:
    * \brief extracts the local function and calls the correct order() method
    * \note  required by `LocalEvaluation::Codim0Interface< ..., 1 >`
    */
-  template< class R, int rT, int rCT >
+  template< class R, size_t rT, size_t rCT >
   size_t order(const LocalfunctionTupleType& localFuncs,
                const Stuff::LocalfunctionSetInterface
                    < EntityType, DomainFieldType, dimDomain, R, rT, rCT >& testBase) const
@@ -94,7 +94,7 @@ public:
    * \brief extracts the local function and calls the correct evaluate() method
    * \note  required by `LocalEvaluation::Codim0Interface< ..., 1 >`
    */
-  template< class R, int rT, int rCT >
+  template< class R, size_t rT, size_t rCT >
   void evaluate(const LocalfunctionTupleType& localFuncs,
                 const Stuff::LocalfunctionSetInterface< EntityType, DomainFieldType, dimDomain, R, rT, rCT >& testBase,
                 const Dune::FieldVector< DomainFieldType, dimDomain >& localPoint,
@@ -110,7 +110,7 @@ public:
    *        - `LocalEvaluation::Codim0Interface< ..., 2 >`
    *        - `LocalEvaluation::Codim1Interface< ..., 2 >`
    */
-  template< class R, int rT, int rCT, int rA, int rCA >
+  template< class R, size_t rT, size_t rCT, size_t rA, size_t rCA >
   size_t order(const LocalfunctionTupleType& localFuncs,
                const Stuff::LocalfunctionSetInterface
                    < EntityType, DomainFieldType, dimDomain, R, rT, rCT >& testBase,
@@ -125,7 +125,7 @@ public:
    * \brief extracts the local function and calls the correct evaluate() method
    * \note  required by `LocalEvaluation::Codim0Interface< ..., 2 >`
    */
-  template< class R, int rT, int rCT, int rA, int rCA >
+  template< class R, size_t rT, size_t rCT, size_t rA, size_t rCA >
   void evaluate(const LocalfunctionTupleType& localFuncs,
                 const Stuff::LocalfunctionSetInterface
                     < EntityType, DomainFieldType, dimDomain, R, rT, rCT >& testBase,
@@ -142,7 +142,7 @@ public:
    * \brief extracts the local function and calls the correct evaluate() method
    * \note  required by `LocalEvaluation::Codim1Interface< ..., 1 >`
    */
-  template< class IntersectionType, class R, int r, int rC >
+  template< class IntersectionType, class R, size_t r, size_t rC >
   void evaluate(const LocalfunctionTupleType& localFuncs,
                 const Stuff::LocalfunctionSetInterface< EntityType, DomainFieldType, dimDomain, R, r, rC >& testBase,
                 const IntersectionType& intersection,
@@ -157,7 +157,7 @@ public:
    * \brief extracts the local function and calls the correct evaluate() method
    * \note  required by `LocalEvaluation::Codim1Interface< ..., 2 >`
    */
-  template< class IntersectionType, class R, int rT, int rCT, int rA, int rCA >
+  template< class IntersectionType, class R, size_t rT, size_t rCT, size_t rA, size_t rCA >
   void evaluate(const LocalfunctionTupleType& localFunctions,
                 const Stuff::LocalfunctionSetInterface
                     < EntityType, DomainFieldType, dimDomain, R, rT, rCT >& testBase,
@@ -184,7 +184,7 @@ public:
   /**
    * \note implementation for `LocalEvaluation::Codim0Interface< ..., 1 >`
    */
-  template< class R, int rL, int rCL, int rT, int rCT >
+  template< class R, size_t rL, size_t rCL, size_t rT, size_t rCT >
   size_t order(const Stuff::LocalfunctionInterface< EntityType, DomainFieldType, dimDomain, R, rL, rCL >& localFunction,
                const Stuff::LocalfunctionSetInterface
                   < EntityType, DomainFieldType, dimDomain, R, rT, rCT >& testBase) const
@@ -221,7 +221,7 @@ public:
    *       - `LocalEvaluation::Codim0Interface< ..., 2 >`
    *       - `LocalEvaluation::Codim1Interface< ..., 2 >`
    */
-  template< class R, int rL, int rCL, int rT, int rCT, int rA, int rCA >
+  template< class R, size_t rL, size_t rCL, size_t rT, size_t rCT, size_t rA, size_t rCA >
   size_t order(const Stuff::LocalfunctionInterface< EntityType, DomainFieldType, dimDomain, R, rL, rCL >& localFunction,
                const Stuff::LocalfunctionSetInterface< EntityType, DomainFieldType, dimDomain, R, rT, rCT >& testBase,
                const Stuff::LocalfunctionSetInterface
@@ -234,7 +234,7 @@ public:
    * \brief Computes a product evaluation for a scalar local function and scalar or vector valued basefunctionsets.
    * \note  implementation for `LocalEvaluation::Codim0Interface< ..., 2 >`
    */
-  template< class R, int r >
+  template< class R, size_t r >
   void evaluate(const Stuff::LocalfunctionInterface< EntityType, DomainFieldType, dimDomain, R, 1, 1 >& localFunction,
                 const Stuff::LocalfunctionSetInterface< EntityType, DomainFieldType, dimDomain, R, r, 1 >& testBase,
                 const Stuff::LocalfunctionSetInterface< EntityType, DomainFieldType, dimDomain, R, r, 1 >& ansatzBase,
@@ -294,7 +294,7 @@ public:
    * \brief Computes a product evaluation for a scalar local function and scalar or vector valued basefunctionsets.
    * \note  implementation for `LocalEvaluation::Codim1Interface< ..., 2 >`
    */
-  template< class IntersectionType, class R, int r >
+  template< class IntersectionType, class R, size_t r >
   void evaluate(const Stuff::LocalfunctionInterface< EntityType, DomainFieldType, dimDomain, R, 1, 1 >& localFunction,
                 const Stuff::LocalfunctionSetInterface< EntityType, DomainFieldType, dimDomain, R, r, 1 >& testBase,
                 const Stuff::LocalfunctionSetInterface< EntityType, DomainFieldType, dimDomain, R, r, 1 >& ansatzBase,
